Create Info Puller for CSGO market
In this form the code will pull the name, quantity and price as well as the thumbnail for the top search and display it using tkinter. This is a very simple, hacked together display method, as I work on it hopefully I'll be able to build something that displays the top ten searches with thumbnails.